#73 [feat] : 모니터링 로직 리팩토링

See merge request s11-final/S11P31A509!100

+++ ttak/app/src/main/java/com/ttak/android/domain/model/FriendStoty.kt
@@ -1,7 +1,7 @@
 package com.ttak.android.domain.model
 
 data class FriendStory(
-    val friendId: Int,
+    val friendId: Long,
     val friendName: String,
     val friendImg: String,
     val status: Int
