+++ ttak/app/src/main/java/com/ttak/android/domain/model/MessageData.kt
@@ -1,6 +1,6 @@
 package com.ttak.android.domain.model
 
 data class MessageData (
-    val userId: Int, // 공격받는 사람 Id
+    val userId: Long, // 공격받는 사람 Id
     val data: String // 메시지 내용
 )
